% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%<>\%}
\alias{\%<>\%}
\title{Assignment pipe operator}
\usage{
lhs \%<>\% rhs
}
\value{
The left hand side object with the right hand side applied to it. Does  reassign this value to the left hand side.
}
\description{
See \code{magrittr::\link[magrittr:compound]{\%<>\%}} for details.
}
\keyword{internal}
