% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{endpoint_input_mapping_nested}
\alias{endpoint_input_mapping_nested}
\title{Nested Endpoint-Input Mapping}
\format{
A tibble with 70 rows and 2 variables:
\describe{
  \item{endpoint}{name of the API endpoint}
  \item{inputs}{a list column containing one or more inputs and a boolean indicating whether they are required for that endpoint. Can be unnested with \code{tidyr::unnest}}
}
}
\source{
\url{http://api.votesmart.org/docs/}
}
\usage{
endpoint_input_mapping_nested
}
\description{
Nested tibble containing the mapping between each endpoint, the inputs it takes, and whether those inputs are required.
}
\keyword{datasets}
