% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_get_measures.R
\name{measure_get_measures}
\alias{measure_get_measures}
\title{Get information on a ballot measure}
\usage{
measure_get_measures(measure_ids, verbose = TRUE)
}
\arguments{
\item{measure_ids}{Vector of ballot measure ids.}

\item{verbose}{Should cases when no data is available be messaged?}
}
\value{
A dataframe with the columns \code{measure_id, measure_code, title, election_date, election_type, outcome, yes_votes, no_votes, summary, summary_url, measure_text, text_url, pro_url, con_url}.
}
\description{
Ballot measure ids can be found with the \code{\link{measure_get_measures_by_year_state}} function.
}
\examples{
\dontrun{
measure_get_measures("1234")
}
}
