% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/office_get_levels.R
\name{office_get_levels}
\alias{office_get_levels}
\title{Get office levels}
\usage{
office_get_levels()
}
\value{
A dataframe with the columns \code{office_level_id} and \code{name}.
}
\description{
These are currently: F for Federal, S for State, and L for Local.
}
\examples{
\dontrun{
office_get_levels()
}
}
