% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/office_get_offices_by_level.R
\name{office_get_offices_by_level}
\alias{office_get_offices_by_level}
\title{Get offices by level}
\usage{
office_get_offices_by_level(office_level_ids)
}
\arguments{
\item{office_level_ids}{Vector of office levels.}
}
\value{
A dataframe with columns \code{office_id, name, title, office_level_id, office_type_id, office_branch_id, short_title}.
}
\description{
Get offices by level
}
\examples{
\dontrun{
office_get_offices_by_level("F")

office_get_levels() \%>\%
  pull(office_level_id) \%>\%
  .[1] \%>\%
  office_get_offices_by_level()
}
}
