% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating_get_categories.R
\name{rating_get_categories}
\alias{rating_get_categories}
\title{Get categories that contain ratings by state}
\usage{
rating_get_categories(state_ids = NA)
}
\arguments{
\item{state_ids}{A vector of state abbreviations. Defaults to \code{NA} for national.}
}
\value{
A dataframe with columns \code{category_id, name, state_id}.
}
\description{
Get categories that contain ratings by state
}
\examples{
\dontrun{
rating_get_categories("NM")
}
}
