% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating_get_sig.R
\name{rating_get_sig}
\alias{rating_get_sig}
\title{Get information on a SIG (Special Interest Group) by its ID}
\usage{
rating_get_sig(sig_ids, verbose = TRUE)
}
\arguments{
\item{sig_ids}{Vector of SIG ids.}

\item{verbose}{Should cases when no data is available be messaged?}
}
\value{
A dataframe with the columns \code{sig_id, name, description, state_id, address, city, state, zip, phone_1, phone_2, fax, email, url, contact_name}.
}
\description{
Get information on a SIG (Special Interest Group) by its ID
}
\examples{
\dontrun{
rating_get_sig_list(2) \%>\%
  dplyr::pull(sig_id) \%>\%
  sample(3) \%>\%
  rating_get_sig()
}
}
