% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votesmart-package.R
\docType{package}
\name{votesmart-package}
\alias{votesmart}
\alias{votesmart-package}
\title{votesmart: Wrapper for the Project VoteSmart API}
\description{
An R interface to the Project VoteSmart API.
}
\author{
\strong{Maintainer}: Amanda Dobbyn \email{amanda@deck.tools}

Authors:
\itemize{
  \item Max Wood \email{max@deck.tools}
}

}
\keyword{internal}
