% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votesmart-package.R
\docType{package}
\name{votesmart-package}
\alias{votesmart}
\alias{votesmart-package}
\title{votesmart: Wrapper for the Project 'VoteSmart' API}
\description{
An R interface to the Project 'VoteSmart'<https://justfacts.votesmart.org/> API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/decktools/votesmart/}
  \item Report bugs at \url{https://github.com/decktools/votesmart/issues/}
}

}
\author{
\strong{Maintainer}: Amanda Dobbyn \email{amanda@deck.tools}

Authors:
\itemize{
  \item Max Wood \email{max@deck.tools}
  \item Alyssa Frazee \email{alyssa@deck.tools}
}

Other contributors:
\itemize{
  \item Deck Technologies \email{info@deck.tools} [copyright holder, funder]
}

}
\keyword{internal}
